Moved read/write to repository class
