+++ spring-application/src/main/kotlin/com/tubefans/gamepicker/services/GoogleSheetsService.kt
@@ -1,10 +1,7 @@
 package com.tubefans.gamepicker.services
 
-import com.google.api.client.googleapis.json.GoogleJsonResponseException
-import com.google.api.services.sheets.v4.Sheets
-import com.google.api.services.sheets.v4.model.ValueRange
 import com.tubefans.gamepicker.cache.GoogleSheetCache.Companion.END_ROW_TITLE
-import org.apache.http.client.HttpResponseException
+import com.tubefans.gamepicker.repositories.GoogleSheetsRepository
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
@@ -15,20 +12,9 @@ import toScore
 class GoogleSheetsService
     @Autowired
     constructor(
-        private val sheets: Sheets,
+        private val googleSheetsRepository: GoogleSheetsRepository,
     ) {
         private val logger = LoggerFactory.getLogger(this::class.java)
-
-        fun getValueRange(
-            id: String,
-            range: String,
-        ): ValueRange =
-            try {
-                sheets.spreadsheets().values()[id, range].execute()
-            } catch (e: HttpResponseException) {
-                logger.error("Failed to get value range from Google API", e)
-                ValueRange()
-            }
 
         /**
          * @param id Google Sheet id
@@ -38,17 +24,13 @@ class GoogleSheetsService
         fun getSheet(
             id: String,
             range: String,
-        ): List<List<Any>> =
-            try {
-                sheets
-                    .spreadsheets()
-                    .values()[id, range]
-                    .execute()
-                    .getValues()
-            } catch (e: HttpResponseException) {
-                logger.error("Failed to get sheet from Google API", e)
-                emptyList()
-            }
+        ): List<List<Any>> = googleSheetsRepository.getSheet(id, range)
+
+        fun writeRange(
+            id: String,
+            range: String,
+            values: List<List<String>>,
+        ) = googleSheetsRepository.writeRange(id, range, values)
 
         /**
          * Helper function to process data from Google sheets
@@ -96,24 +78,5 @@ class GoogleSheetsService
             }
 
             return scoreMap
-        }
-
-        fun writeRange(
-            id: String,
-            range: String,
-            values: List<List<String>>,
-        ) {
-            logger.info("Attempting to write to sheet at range {} for values {}", range, values.joinToString { it.joinToString() })
-            val body = ValueRange().setValues(values)
-            try {
-                sheets
-                    .spreadsheets()
-                    .values()
-                    .update(id, range, body)
-                    .setValueInputOption("USER_ENTERED")
-                    .execute()
-            } catch (e: GoogleJsonResponseException) {
-                logger.error("Failed to write to google sheet", e)
-            }
         }
     }
